Refactor: componentize Card
